Consistency in search for class loader
I'm building a system for loading BEAM tasks as plugins and I'm seeing failures when using the Dataflow runner that trace back to this line. Not sure if my fix is the right one -- maybe I'm thinking about this wrong? -- but it stands to reason that the existing code is relying on one of two assumptions: (a) the BEAM classes having been loaded in a custom context class loader or (b) the system class loader being a implemented with `URLClassLoader`. The latter is no longer the case in more recent JVM releases.